+++ CHANGELOG.md
@@ -22,6 +22,10 @@ As a minor extension, we have adopted a slightly different versioning convention
 
 - Add a one line shell installation script for the Mithril nodes.
 
+- **UNSTABLE** Cardano database incremental certification:
+
+  - Implement the new signed entity type `CardanoDatabase`.
+
 - Crates versions:
 
 | Crate | Version |